+++ pkg/sheets/sheets_test.go
@@ -1,4 +1,4 @@
-package sheet
+package sheets
 
 import (
 	"github.com/stretchr/testify/assert"
